[General] UX improvements in loading screens
